fastapi_2/requirements.txt
fastapi
tensorflow
Pillow
numpy
python-multipart
uvicorn